+++ app/src/main/java/com/example/planthelper/data/datasource/local/dao/base/flow/FlowByIdDao.kt
@@ -5,6 +5,6 @@ import kotlinx.coroutines.flow.Flow
 interface FlowByIdDao<T> {
 
     @JvmSuppressWildcards
-    fun flow(id: Int): Flow<T>
+    fun flow(id: Long): Flow<T>
 
 }
