+++ app/src/main/java/com/example/planthelper/data/datasource/local/dao/base/get/GetByIdDao.kt
@@ -3,6 +3,6 @@ package com.example.planthelper.data.datasource.local.dao.base.get
 interface GetByIdDao<T> {
 
     @JvmSuppressWildcards
-    suspend fun get(id: Int): T
+    suspend fun get(id: Long): T
 
 }
